requirements.txt
discord.py
asyncio
PyNaCl
nextcord
wavelink
datetime